Declare variables outside the scope of for loops
This has a slight performance boost since we can eliminate the need of recreating a variable within the scope of a loop multiple times. Instead we just declare the variable outside the scope of the loop one time and allow the loop to peak its performance by skipping the need to recreate that variable with each return. A lot of benchmark testing has shown me that this is by far the fastest way to loop. Using reverse loops or classical loops seem to have marginal differences that are mostly naked to the eye in my opinion. In most cases I have resorted to using reverse loops with a couple of exceptions.